Implement basic logging macros
