+++ src/logging.c
@@ -1,0 +1,18 @@
+#include "logging.h"
+
+/**
+ * @brief Make `printf` send characters through the USART.
+ */
+int _write(int file, char *ptr, int len)
+{
+	int i;
+
+	if (file == 1) {
+		for (i = 0; i < len; i++)
+			usart_send_blocking(USART3, ptr[i]);
+		return i;
+	}
+
+	errno = EIO;
+	return -1;
+}
